Current TMNT Arena engine source code!

+++ trunk/engine/code/bspc/aas_cfg.c
@@ -29,9 +29,9 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "../botlib/l_struct.h"
 #include "../botlib/l_libvar.h"
 
-//structure field offsets
-#define BBOX_OFS(x) (int)&(((aas_bbox_t *)0)->x)
-#define CFG_OFS(x) (int)&(((cfg_t *)0)->x)
+//structure field offsets Turtle Man: int to size_t
+#define BBOX_OFS(x) (size_t)&(((aas_bbox_t *)0)->x)
+#define CFG_OFS(x) (size_t)&(((cfg_t *)0)->x)
 
 //bounding box definition
 fielddef_t bbox_fields[] =
